requirements.txt
numpy==1.23.5
streamlit==1.29.0
Pillow==9.5.0
opencv-python-headless>=4.6.0.66
rembg==2.0.35
keras==2.12.0
torchvision==0.14.1
tensorflow==2.12.0
torch==1.13.1
